requirements.txt
pandas
pymysql
selenium
cryptography
webdriver-manager
google-generativeai